requirements.txt
requests~=2.31.0
pandas
sqlalchemy~=2.0.27
psycopg2-binary
python-dotenv~=1.0.1
pytz~=2024.1
openpyxl
pyodbc